src/requirements.txt
Flask==0.12.2
flask-restful
google-cloud-datastore
requests==2.8.1
